@@ -5,7 +5,10 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using ProAgil.WebAPI.Data;
+using ProAgil.Repository.Context;
+using ProAgil.Repository.Interface;
+using ProAgil.Repository;
+using AutoMapper;
 
 namespace ProAgil.WebAPI
 {
@@ -20,14 +23,29 @@ namespace ProAgil.WebAPI
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllers();
-             services.AddDbContext<DataContext>(options =>
+            services.AddDbContext<ProAgilContext>(options =>
                         options.UseSqlServer(Configuration.GetConnectionString("DefaultConnection")));
-            services.AddCors();                        
+
+            services.AddScoped<IProAgilRepository, ProAgilRepository>();
+
+            services.AddControllers();
+            services.AddControllers().AddNewtonsoftJson();
+                                 
+            services.AddSwaggerGen();   
+            services.AddCors();     
+          
+            services.AddAutoMapper(typeof(Startup));
+             
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
+            app.UseSwagger();
+            app.UseSwaggerUI(c =>
+            {
+                c.SwaggerEndpoint("/swagger/v1/swagger.json", "My API V1");
+            });
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
@@ -35,8 +53,9 @@ namespace ProAgil.WebAPI
 
            // app.UseHttpsRedirection();
             app.UseCors(x => x.AllowAnyOrigin().AllowAnyMethod().AllowAnyHeader());
-            app.UseRouting();
+            app.UseStaticFiles();
 
+            app.UseRouting();
             app.UseAuthorization();
 
             app.UseEndpoints(endpoints =>
